+++ Sanctuary.Census.Common/Objects/Collections/FacilityInfo.cs
@@ -1,5 +1,6 @@
 ﻿using Sanctuary.Census.Common.Attributes;
 using Sanctuary.Census.Common.Objects.CommonModels;
+using System.ComponentModel;
 using System.ComponentModel.DataAnnotations;
 
 namespace Sanctuary.Census.Common.Objects.Collections;
@@ -15,6 +16,7 @@ namespace Sanctuary.Census.Common.Objects.Collections;
 /// <param name="LocationY">The Y coordinate of the facility.</param>
 /// <param name="LocationZ">The Z coordinate of the facility.</param>
 [Collection]
+[Description("Information about in-game facilities. While missing some of the fields of map_region, this collection is guaranteed to be up-to-date")]
 public record FacilityInfo
 (
     [property: Key] ushort ZoneID,