+++ Sanctuary.Census.Common/Objects/Collections/FacilityLink.cs
@@ -1,4 +1,5 @@
 ﻿using Sanctuary.Census.Common.Attributes;
+using System.ComponentModel;
 using System.ComponentModel.DataAnnotations;
 
 namespace Sanctuary.Census.Common.Objects.Collections;
@@ -11,6 +12,7 @@ namespace Sanctuary.Census.Common.Objects.Collections;
 /// <param name="FacilityIdB">The second facility involved in the link.</param>
 /// <param name="Description">The description of the link.</param>
 [Collection]
+[Description("Information about lattice links between facilities. Note that this collection is not automatically updated at this time")]
 public record FacilityLink
 (
     [property:Key] uint ZoneID,