+++ Sanctuary.Census.Common/Objects/Collections/LoadoutSlot.cs
@@ -1,5 +1,6 @@
 ﻿using Sanctuary.Census.Common.Attributes;
 using Sanctuary.Census.Common.Objects.CommonModels;
+using System.ComponentModel;
 using System.ComponentModel.DataAnnotations;
 
 namespace Sanctuary.Census.Common.Objects.Collections;
@@ -16,6 +17,7 @@ namespace Sanctuary.Census.Common.Objects.Collections;
 /// <param name="IsVisible">Whether this slot is visible to the user.</param>
 /// <param name="UiTag">The UI section that this slot appears under.</param>
 [Collection]
+[Description("Represents a slot of an infantry loadout; i.e. the primary weapon or suit slot")]
 public record LoadoutSlot
 (
     [property: Key] uint LoadoutID,